refactor: decouple question repository
