+++ src/libs/modules/database/repositories/question/question.interface.ts
@@ -1,0 +1,8 @@
+export interface QuestionRepository {
+  findFirst;
+  update;
+  create;
+  delete;
+}
+
+export const QuestionRepository = Symbol('QuestionRepository');
